% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mitigations}
\alias{check_mitigations}
\title{Check mitigations file for problems}
\usage{
check_mitigations(mitigations, pathways = "data/poe.json")
}
\arguments{
\item{pathways}{}
}
\description{
Check mitigations file for problems
}
\examples{
mitigations <- readxl::read_excel("data/mitigations.xlsx")

check_mitigations(mitigations)

m <- data.frame(start_node = c(1, 1), end_node = c(99, 99), edge = c(NA, 5))
check_mitigations(m)

}
