% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_report}
\alias{create_report}
\title{Create html report of diagram}
\usage{
create_report(
  vc,
  a,
  m = NULL,
  m_df = NULL,
  notes = NULL,
  lang = lang,
  path = paste0("report_", Sys.Date(), ".html")
)
}
\description{
Create html report of diagram
}
\examples{
v <- "Terrestrial and Semi-Aquatic SAR"
a <- c("Shoreline / Bank stabilization", "Water extraction")
m <- c("terrestrial_and_semi-aquatic_sar_relocate", "terrestrial_and_semi-aquatic_sar_create_habitat")
m_df <- read_mitigations()
create_report(v, a, m, m_df, lang = "en")
create_report(v, a, m, m_df, lang = "fr")

# No Mitigations
create_report(
  "Marine Birds",
  "Aircraft overflights / runway activity",
  lang = "en"
)

# Mitigations
create_report(
  vc = "Marine Birds",
  a = "Aircraft overflights / runway activity",
  m = "Mitigation",
  m_df = data.frame(
    start_node = NA,
    end_node = NA,
    edge = 67,
    m_id = "vc_mitigation",
    short = "Mitigation",
    long = "Test"
  ),
  lang = "en"
)
}
