% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_flowchart}
\alias{make_flowchart}
\title{[ON HOLD] Create Diagrammr mermaid flowchart}
\usage{
make_flowchart(pathway)
}
\arguments{
\item{pathway}{List. Current set of diagram pathways for a single component.}
}
\value{
DiagrammeR htmlwidget.
}
\description{
[ON HOLD] Create Diagrammr mermaid flowchart
}
