% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_visnetwork}
\alias{make_visnetwork}
\title{Create visnetwork interactive figure}
\usage{
make_visnetwork(pathway, width = NULL, height = NULL)
}
\arguments{
\item{pathway}{List. Current set of diagram pathways for a single component.}
}
\value{
visnetwork diagram
}
\description{
Create visnetwork interactive figure
}
\examples{
pathways <- jsonlite::read_json("data/poe.json", simplifyVector = FALSE)
ref <- as.data.frame(readRDS("data/act2Pres.rds"))

pathway <- prep_pathways(
  pathways,
  ref,
  vc = "Terrestrial and Semi-Aquatic SAR",
  a = "Shoreline / Bank stabilization"
)

make_visnetwork(pathway)
}
