% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{poe_app}
\alias{poe_app}
\title{Launch the Pathways of Effect Shiny App}
\usage{
poe_app()
}
\description{
Launches the app. Note that the App data must be in the current working
directory for the app to work.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
poe_app()
\dontshow{\}) # examplesIf}
}
