% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_pathways}
\alias{prep_pathways}
\title{Prepare pruned pathways from spreadsheet}
\usage{
prep_pathways(pathways, ref, vc, a, lang = "en")
}
\arguments{
\item{pathways}{Data frame. Pathway diagrams loaded from `extdata/pathways`}

\item{ref}{Data.frame. References data from `data/act2Pres.rds` including
Valued Components, Activities and Stressors.}

\item{vc}{Character. Valued Component to select pathways for}

\item{a}{Character. Activity to select pathways for}

\item{lang}{Character. Language for display, "en" (English) or "fr" (French).}
}
\description{
Based on full pathway diagrams, prune according to selections
}
\examples{
ref <- readRDS("inst/extdata/act2Pres.rds") |>  as.data.frame()

prep_pathways(
  read_pathways(),
  ref,
  vc = "Terrestrial and Semi-Aquatic SAR",
  a = "Shoreline / Bank stabilization"
)
}
