% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_components}
\alias{read_components}
\title{Read components data}
\usage{
read_components(file_name = "components.xlsx", dir = NULL)
}
\arguments{
\item{file_name}{Character. Name of Excel file to read. Defaults to
"components.xlsx".}

\item{dir}{Character. Directory containing the file. Defaults to `NULL`
which uses package data or looks for the file in the current working
directory.}
}
\value{
Data frame with columns:
* `valued_component` - Valued component name
* `activities` - Activity name
* `stressors` - Stressor name
}
\description{
Reads components Excel file containing the mapping among valued components,
activities, and stressors. Used to create the stressors checkbox selections
and to select which nodes are included in the diagram.
}
\examples{
\dontshow{if (have_data()) withAutoprint(\{ # examplesIf}
read_components()
\dontshow{\}) # examplesIf}
}
