% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_mitigations}
\alias{read_mitigations}
\title{Read mitigations data}
\usage{
read_mitigations(file_name = "mitigations.xlsx", dir = NULL)
}
\arguments{
\item{file_name}{Character. Name of Excel file to read. Defaults to
"mitigations.xlsx".}

\item{dir}{Character. Directory containing the file. Defaults to `NULL`
which uses package data or looks for the file in the current working
directory.}
}
\value{
Data frame with columns:
* `valued_component` - Valued component name
* `start_node` - Starting node ID for mitigation edge
* `end_node` - Ending node ID for mitigation edge
* `short` - Short mitigation description
* `long` - Long mitigation description
* `edge_id` - Edge identifier (created from start and end nodes)
* `m_id` - Mitigation identifier (created from VC and short description)
}
\description{
Reads mitigations Excel file containing mitigation descriptions and the
edges they apply to. Creates edge IDs and mitigation IDs. Used to create
mitigations checkboxes and to apply mitigations to the diagrams.
}
\examples{
\dontshow{if (have_data()) withAutoprint(\{ # examplesIf}
read_mitigations()
\dontshow{\}) # examplesIf}
}
