% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_sectors}
\alias{read_sectors}
\title{Read activities data}
\usage{
read_sectors(file_name = "sectors.xlsx", dir = NULL)
}
\arguments{
\item{file_name}{Character. Name of Excel file to read. Defaults to
"sectors.xlsx".}

\item{dir}{Character. Directory containing the file. Defaults to `NULL`
which uses package data or looks for the file in the current working
directory.}
}
\value{
Data frame with columns:
* `activities` - Activity name
* `sector` - Sector name
}
\description{
Reads activities Excel file containing activity names organized by sector.
Used to create Sector drop down to pre-select activities.
}
\examples{
\dontshow{if (have_data()) withAutoprint(\{ # examplesIf}
read_sectors()
\dontshow{\}) # examplesIf}
}
