% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_tables}
\alias{report_tables}
\title{Create list of tables for reports}
\usage{
report_tables(tbl, cols, lang)
}
\arguments{
\item{tbl}{Data frame. Table data to format.}

\item{cols}{Named character vector. Columns to keep in table. Names become
column headers.}

\item{lang}{Character. Language code ("en" or "fr") for translations.}
}
\value{
Named list of reactable tables, one per unique value in first
  column.
}
\description{
Creates list of reactable tables from data frame, split by first column and
with translated labels.
}
