% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_tables_print}
\alias{report_tables_print}
\title{Print a list of reactable tables in report}
\usage{
report_tables_print(tbls)
}
\arguments{
\item{tbls}{List of reactable tables created by `report_tables()`}
}
\description{
Print a list of reactable tables in report
}
