% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_mitigation}
\alias{add_mitigation}
\title{Add mitigations to pathways}
\usage{
add_mitigation(pathway, mitigations, m, lang = "en")
}
\arguments{
\item{pathway}{List. Current set of diagram pathways for a single component.}

\item{mitigations}{Data frame. Master list of mitigation metadata.}

\item{m}{Character vector. Currently selected mitigations (refer to
`short_en` in `mitigations`) to add to the pathway.}

\item{lang}{Character. Language for display, "en" (English) or "fr" (French).}
}
\description{
Disables nodes and edges downstream of a mitigation (disabled edge) which are
not maintained by alternate pathways.
}
\examples{
pathways <- jsonlite::read_json(
 path = "data/poe.json",
 simplifyVector = FALSE
)
mitigations <- readxl::read_excel("data/mitigations.xlsx")
ref <- readRDS("data/act2Pres.rds") |>
  as.data.frame()

p <- prep_pathways(
  pathways,
  ref,
  vc = "Terrestrial and Semi-Aquatic SAR",
  a = "Shoreline / Bank stabilization"
)

make_visnetwork(p)
p1 <- add_mitigation(p, mitigations, "Selective work")

make_visnetwork(p1)
make_flowchart(p1)
}
