% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mitigations}
\alias{check_mitigations}
\title{Check mitigations file for problems}
\usage{
check_mitigations(mitigations = read_mitigations(), pathways = read_pathways())
}
\arguments{
\item{mitigations}{Data frame. Mitigations data from [read_mitigations()].}

\item{pathways}{Data frame. Pathways data from [read_pathways()].}
}
\value{
List with validation status (invisibly). Stops with error if
  problems are found.
}
\description{
Validates that all mitigation edges exist in the pathway data.
}
\examples{
\dontshow{if (have_data()) withAutoprint(\{ # examplesIf}
check_mitigations()
\dontshow{\}) # examplesIf}
}
