% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mitigations}
\alias{check_mitigations}
\title{Check mitigations file for problems}
\usage{
check_mitigations(mitigations = read_mitigations(), pathways = read_pathways())
}
\arguments{
\item{pathways}{}
}
\description{
Check mitigations file for problems
}
\examples{
check_mitigations()
}
