% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_report}
\alias{create_report}
\title{Create html report of diagram}
\usage{
create_report(
  vc,
  a,
  m = NULL,
  m_df = NULL,
  notes = NULL,
  lang = lang,
  path = paste0("report_", Sys.Date(), ".html")
)
}
\description{
Create html report of diagram
}
\examples{
pathways <- jsonlite::read_json(
 path = "data/poe.json",
 simplifyVector = FALSE
)

mitigations <- readxl::read_excel("data/mitigations.xlsx")
ref <- readRDS("data/act2Pres.rds") |>
  as.data.frame()


create_report(
  "Terrestrial and Semi-Aquatic SAR",
  c(
    "Shoreline / Bank stabilization",
    "Land and Vegetation clearing and vegetation maintenance"
  ),
  c("Relocate", "Create Habitat"),
  m_df = mitigations,
  lang = "en"
)

# No Mitigations
create_report(
  "Marine Birds",
  "Aircraft overflights / runway activity",
  lang = "en"
)

# Mitigations
create_report(
  vc = "Marine Birds",
  a = "Aircraft overflights / runway activity",
  m = "Mitigation",
  m_df = data.frame(
    start_node = NA,
    end_node = NA,
    edge = 67,
    short_en = "Mitigation",
    long_en = "Test"
  ),
  lang = "en"
)
}
