% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_report}
\alias{create_report}
\title{Create html report of diagram}
\usage{
create_report(
  vc,
  a,
  m = NULL,
  m_df = read_mitigations(),
  notes = NULL,
  project_name = "Project Placeholder",
  project_status = "Status Placeholder",
  lang = lang,
  path = paste0("report_", Sys.Date(), ".html")
)
}
\arguments{
\item{vc}{Character. Valued component to include in report.}

\item{a}{Character vector. Activities to include in report.}

\item{m}{Character vector. Mitigation IDs to include. Defaults to `NULL`.}

\item{m_df}{Data frame. Mitigation metadata. Defaults to [read_mitigations()]}

\item{notes}{Character. User notes to include in report. Defaults to `NULL`.}

\item{project_name}{Character. Project name for report header. Defaults to
"Project Placeholder".}

\item{project_status}{Character. Project status for report header. Defaults
to "Status Placeholder".}

\item{lang}{Character. Language code ("en" or "fr").}

\item{path}{Character. Output file path. Defaults to
`paste0("report_", Sys.Date(), ".html")`.}
}
\value{
Character. Path to created HTML report file.
}
\description{
Generates an HTML report showing the pathway diagram with selected
mitigations and project information.
}
\examples{
\dontshow{if (have_data()) withAutoprint(\{ # examplesIf}
v <- "Terrestrial and Semi-Aquatic SAR"
a <- c("Shoreline / Bank stabilization", "Water extraction")
m <- c("terrestrial_and_semi-aquatic_sar_relocate",
       "terrestrial_and_semi-aquatic_sar_create_habitat")
m_df <- read_mitigations()
create_report(v, a, m, m_df, lang = "en")
create_report(v, a, m, m_df, lang = "fr")

# No Mitigations
v <- "Marine Birds"
a <- "Aircraft overflights / runway activity"
create_report(v, a, lang = "en")
create_report(v, a, lang = "fr")

# Cleanup
file.remove(paste0("report_", Sys.Date(), ".html"))
\dontshow{\}) # examplesIf}
}
