% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_flowchart}
\alias{make_flowchart}
\title{Create Diagrammr mermaid flowchart}
\usage{
make_flowchart(pathway)
}
\arguments{
\item{pathway}{List. Current set of diagram pathways for a single component.}
}
\description{
Create Diagrammr mermaid flowchart
}
\examples{
pathways <- jsonlite::read_json("data/poe.json", simplifyVector = FALSE)
ref <- as.data.frame(readRDS("data/act2Pres.rds"))

pathway <- prep_pathways(
  pathways,
  ref,
  vc = "Terrestrial and Semi-Aquatic SAR",
  a = "Shoreline / Bank stabilization"
)

make_flowchart(pathway)
}
