% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe-package.R
\docType{package}
\name{poe-package}
\alias{poe}
\alias{poe-package}
\title{poe: CWS Shiny Explorer for Pathways of Effects Diagram models}
\description{
Interactive explorer for Pathways of Effects diagram models developed by CWS to visually represent cause and effect relationships between project develpments and environmental components.
}
\author{
\strong{Maintainer}: Steffi LaZerte \email{sel@steffilazerte.ca}

Authors:
\itemize{
  \item Thomas Roh \email{thomas.roh@delveds.com}
}

Other contributors:
\itemize{
  \item Canadian Wildlife Service [copyright holder]
}

}
\keyword{internal}
