% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_pathways}
\alias{prep_pathways}
\title{Prepare pruned pathways from spreadsheet}
\usage{
prep_pathways(pathways, ref, vc, a, lang = "en")
}
\arguments{
\item{pathways}{Data frame. Pathway diagrams from [read_pathways()].}

\item{ref}{Data frame. Component reference data from [read_components()]
including valued components, activities, and stressors.}

\item{vc}{Character. Valued component to select pathways for.}

\item{a}{Character. Activity to select pathways for.}

\item{lang}{Character. Language for display, "en" (English) or "fr" (French).
Defaults to "en".}
}
\value{
Named list with:
* `nodes` - Data frame of nodes in the pruned pathway
* `edges` - Data frame of edges in the pruned pathway
}
\description{
Based on full pathway diagrams, prune according to selections of valued
component and activity to show only relevant pathways and stressors.
}
\examples{
\dontshow{if (have_data()) withAutoprint(\{ # examplesIf}
prep_pathways(
  read_pathways(),
  read_components(),
  vc = "Terrestrial and Semi-Aquatic SAR",
  a = "Shoreline / Bank stabilization"
)
\dontshow{\}) # examplesIf}
}
