% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_tables}
\alias{report_tables}
\title{Create list of tables for reports}
\usage{
report_tables(tbl, cols, lang)
}
\arguments{
\item{tbl}{Data frame.}

\item{cols}{Named character vector. Columns to keep in table, names become
pretty column names.}

\item{lang}{Character. 'en' or 'fr' language to translate values to.}
}
\value{
List of reactable tables
}
\description{
Create list of tables for reports
}
