% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_tables_print}
\alias{report_tables_print}
\title{Print a list of reactable tables in report}
\usage{
report_tables_print(tbls)
}
\arguments{
\item{tbls}{Named list. List of reactable tables created by
[report_tables()].}
}
\value{
NULL (invisibly). Prints tables to output.
}
\description{
Prints reactable tables with headers for use in Quarto reports. This
function is used internally by the report template and is not intended to be
called directly by the user.
}
