% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{translate_text}
\alias{translate_text}
\title{Translates text from english to french}
\usage{
translate_text(x, lang = "en", dict = NULL)
}
\arguments{
\item{x}{Character. English text to translate}

\item{lang}{Character. Language for display, "en" (English) or "fr" (French).}

\item{translations}{Data.frame. with columns "en" and "fr" for translations}
}
\description{
Translates text from english to french
}
\examples{
enFr <- read.csv("data/en-fr-table.csv")
translate_text("Increase in Edge Habitat", "fr", dict = enFr)
}
